Fixed subtexture width and height not calculating correctly

+++ Daybreak/src/Daybreak/Renderer/SubTexture.h
@@ -13,13 +13,13 @@ namespace Daybreak
 
         const glm::vec2* GetTextureCoords() const { return m_TexCoords; }
         const Ref<Texture2D> GetTexture() const { return m_Texture; }
-        const float GetWidth() const { return m_Width; }
-        const float GetHeight() const { return m_Height; }
+        const uint32_t GetWidth() const { return m_Width; }
+        const uint32_t GetHeight() const { return m_Height; }
 
         static Ref<SubTexture2D> Create(const Ref<Texture2D>& texture, const glm::vec2& coords, const glm::vec2& cellSize, const glm::vec2& cellCount = { 1, 1 });
     private:
         Ref<Texture2D> m_Texture;
         glm::vec2 m_TexCoords[4];
-        float m_Width, m_Height;
+        uint32_t m_Width, m_Height;
     };
 }
